applications/basic_py_sql_app/ch10/requirements.txt
Flask>=3.0,<4
Flask-WTF>=1.2,<2
WTForms>=3.1,<4
mongoengine>=0.29.0,<0.30
pymongo>=4.6,<5
gunicorn>=22.0.0
click>=8.1.0
python-dotenv>=1.0.0
